Extract input validation logic.
